+++ CS201_Project/src/MusicPlayer.java
@@ -354,31 +354,39 @@ public class MusicPlayer extends JPanel{
 				if (favoriteLabel.getIcon() == emptyHeart)
 				{
 					favoriteLabel.setIcon(fullHeart);
-					try
+					if (!musicObject.getFavoritedBool())
 					{
-						PreparedStatement ps = (PreparedStatement) ConnectionClass.conn.prepareStatement("INSERT INTO favorite_songs (user_id, song_id)" + "VALUES (?, ?)");
-						ps.setInt(1, LoggedInDriverGUI.userID);
-						ps.setInt(2, musicObject.getMusicID());
-						ps.executeUpdate();
-						ps.close();
-					} catch (SQLException e1)
-					{
-						e1.printStackTrace();
+						try
+						{
+							PreparedStatement ps = (PreparedStatement) ConnectionClass.conn.prepareStatement("INSERT INTO favorite_songs (user_id, song_id)" + "VALUES (?, ?)");
+							ps.setInt(1, LoggedInDriverGUI.userID);
+							ps.setInt(2, musicObject.getMusicID());
+							ps.executeUpdate();
+							ps.close();
+						} catch (SQLException e1)
+						{
+							e1.printStackTrace();
+						}
+						musicObject.setFavoritedBool(true);
 					}
 				}
 				else
 				{
 					favoriteLabel.setIcon(emptyHeart);
-					try
+					if (musicObject.getFavoritedBool())
 					{
-						PreparedStatement ps = (PreparedStatement) ConnectionClass.conn.prepareStatement("DELETE FROM favorite_songs WHERE " + "user_id = ?" + " and " + "song_id = ?");
-						ps.setInt(1, LoggedInDriverGUI.userID);
-						ps.setInt(2, musicObject.getMusicID());
-						ps.executeUpdate();
-						ps.close();
-					} catch (SQLException e1)
-					{
-						e1.printStackTrace();
+						try
+						{
+							PreparedStatement ps = (PreparedStatement) ConnectionClass.conn.prepareStatement("DELETE FROM favorite_songs WHERE " + "user_id = ?" + " and " + "song_id = ?");
+							ps.setInt(1, LoggedInDriverGUI.userID);
+							ps.setInt(2, musicObject.getMusicID());
+							ps.executeUpdate();
+							ps.close();
+						} catch (SQLException e1)
+						{
+							e1.printStackTrace();
+						}
+						musicObject.setFavoritedBool(false);
 					}
 				}
